reorganize boostPython: split files and reduce compilation time
