+++ boostPython/ScalarFieldPy.hpp
@@ -1,0 +1,24 @@
+//##########################################################################
+//#                                                                        #
+//#                                PYCC                                    #
+//#                                                                        #
+//#  This program is free software; you can redistribute it and/or modify  #
+//#  it under the terms of the GNU Library General Public License as       #
+//#  published by the Free Software Foundation; version 2 or later of the  #
+//#  License.                                                              #
+//#                                                                        #
+//#  This program is distributed in the hope that it will be useful,       #
+//#  but WITHOUT ANY WARRANTY; without even the implied warranty of        #
+//#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the          #
+//#  GNU General Public License for more details.                          #
+//#                                                                        #
+//#          Copyright 2021 Paul RASCLE www.openfields.fr                  #
+//#                                                                        #
+//##########################################################################
+
+#ifndef SCALARFIELDPY_HPP_
+#define SCALARFIELDPY_HPP_
+
+void export_ScalarField();
+
+#endif
